Traverse DLs and child DLs to see if an account is allowed to send to this DL
Used to work for a large (40k+) company with tens of thousands of distribution lists, many of which have parent/child structure, and almost all are restricted, i.e., a list of users/groups are defined for the allowed sender list.  You can just imagine how tedious it would be for someone to verify if a person (or service account) is allowed to send emails to that DL and check the (potentially) hundreds of child-DLs underneath it.